+++ actix-identity/CHANGES.md
@@ -2,7 +2,7 @@
 
 ## Unreleased - 2022-xx-xx
 
-- Replace use of anyhow::Error with identity specific errors. [#296]
+- Replace use of `anyhow::Error` with specific error types. [#296]
 - Minimum supported Rust version (MSRV) is now 1.59 due to transitive `time` dependency.
 
 [#296]: https://github.com/actix/actix-extras/pull/296