+++ actix-identity/CHANGES.md
@@ -3,7 +3,7 @@
 ## Unreleased - 2022-xx-xx
 
 - Replace use of `anyhow::Error` with specific error types. [#296]
-- Minimum supported Rust version (MSRV) is now 1.60.
+- Minimum supported Rust version (MSRV) is now 1.68.
 
 [#296]: https://github.com/actix/actix-extras/pull/296